feat(DTFS2-7498): refactor amendments E2E tests
