+++ e2e-tests/gef/cypress/e2e/pages/amendments/do-you-have-a-facility-end-date.js
@@ -5,6 +5,7 @@ const doYouHaveAFacilityEndDate = {
   inlineError: () => cy.get('[data-cy="is-using-facility-end-date-error"]'),
   pageHeading: () => cy.get('[data-cy="page-heading"]'),
   backLink: () => cy.get('[data-cy="back-link"]'),
+  cancelLink: () => cy.get('[data-cy="cancel-link"]'),
 };
 
 module.exports = doYouHaveAFacilityEndDate;
