+++ e2e-tests/gef/cypress/e2e/pages/amendments/eligibility.js
@@ -8,6 +8,7 @@ const eligibility = {
   criterionRadiosText: (id) => cy.get(`[data-cy="radio-wrapper-${id}"]`),
   pageHeading: () => cy.get('[data-cy="page-heading"]'),
   backLink: () => cy.get('[data-cy="back-link"]'),
+  cancelLink: () => cy.get('[data-cy="cancel-link"]'),
 };
 
 module.exports = eligibility;
