+++ e2e-tests/ukef/cypress/e2e/journeys/portal/amendments/pages/cover-end-date.spec.js
@@ -1,17 +1,17 @@
 import { sub } from 'date-fns';
 import { now } from '@ukef/dtfs2-common';
-import relative from '../../../relativeURL';
-import MOCK_USERS from '../../../../../../e2e-fixtures/portal-users.fixture';
-import { MOCK_APPLICATION_AIN_DRAFT } from '../../../../../../e2e-fixtures/gef/mocks/mock-deals';
-import { anIssuedCashFacility } from '../../../../../../e2e-fixtures/mock-gef-facilities';
-import { applicationPreview } from '../../../../../../gef/cypress/e2e/pages';
-import whatDoYouNeedToChange from '../../../../../../gef/cypress/e2e/pages/amendments/what-do-you-need-to-change';
-import coverEndDate from '../../../../../../gef/cypress/e2e/pages/amendments/cover-end-date';
-import { sixYearsOneDay } from '../../../../../../e2e-fixtures/dateConstants';
+import relative from '../../../../relativeURL';
+import MOCK_USERS from '../../../../../../../e2e-fixtures/portal-users.fixture';
+import { MOCK_APPLICATION_AIN_DRAFT } from '../../../../../../../e2e-fixtures/gef/mocks/mock-deals';
+import { anIssuedCashFacility } from '../../../../../../../e2e-fixtures/mock-gef-facilities';
+import { applicationPreview } from '../../../../../../../gef/cypress/e2e/pages';
+import whatDoYouNeedToChange from '../../../../../../../gef/cypress/e2e/pages/amendments/what-do-you-need-to-change';
+import coverEndDate from '../../../../../../../gef/cypress/e2e/pages/amendments/cover-end-date';
+import { sixYearsOneDay } from '../../../../../../../e2e-fixtures/dateConstants';
 
 const { BANK1_MAKER1 } = MOCK_USERS;
 
-context('Change cover end date journey - unhappy path', () => {
+context('Amendments - Cover end date - page tests', () => {
   /**
    * @type {string}
    */
@@ -55,6 +55,9 @@ context('Change cover end date journey - unhappy path', () => {
 
           amendmentId = urlSplit[9];
         });
+
+        whatDoYouNeedToChange.coverEndDateCheckbox().click();
+        cy.clickContinueButton();
       });
     });
   });
@@ -67,27 +70,10 @@ context('Change cover end date journey - unhappy path', () => {
   beforeEach(() => {
     cy.clearSessionCookies();
     cy.login(BANK1_MAKER1);
-  });
-
-  it('should render an error if nothing is selected to change', () => {
-    cy.visit(relative(`/gef/application-details/${dealId}/facilities/${facilityId}/amendments/${amendmentId}/what-do-you-need-to-change`));
-
-    cy.clickContinueButton();
-
-    whatDoYouNeedToChange.errorSummary().should('be.visible');
-    whatDoYouNeedToChange.errorSummary().contains('Select if you need to change the facility cover end date, value or both');
-
-    whatDoYouNeedToChange.amendmentOptionsInlineError().should('be.visible');
-    whatDoYouNeedToChange.amendmentOptionsInlineError().contains('Select if you need to change the facility cover end date, value or both');
+    cy.visit(relative(`/gef/application-details/${dealId}/facilities/${facilityId}/amendments/${amendmentId}/cover-end-date`));
   });
 
   it('should render an error if no cover end date is provided', () => {
-    cy.visit(relative(`/gef/application-details/${dealId}/facilities/${facilityId}/amendments/${amendmentId}/what-do-you-need-to-change`));
-
-    whatDoYouNeedToChange.coverEndDateCheckbox().click();
-    cy.clickContinueButton();
-
-    cy.url().should('eq', relative(`/gef/application-details/${dealId}/facilities/${facilityId}/amendments/${amendmentId}/cover-end-date`));
     cy.clickContinueButton();
 
     coverEndDate.errorSummary().should('be.visible');
@@ -128,5 +114,11 @@ context('Change cover end date journey - unhappy path', () => {
     const expectedErrorMessage = 'The new cover end date cannot be greater than 6 years in the future';
     coverEndDate.coverEndDateInlineError().should('be.visible');
     coverEndDate.coverEndDateInlineError().contains(expectedErrorMessage);
+  });
+
+  it('should navigate to cancel page when cancel is clicked', () => {
+    coverEndDate.cancelLink().click();
+
+    cy.url().should('eq', relative(`/gef/application-details/${dealId}/facilities/${facilityId}/amendments/${amendmentId}/cancel`));
   });
 });
